backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.25.2
openai==1.3.7
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
reportlab==4.0.7
weasyprint==60.2
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
pydantic==2.5.0
pydantic[email]==2.5.0
httpx==0.25.2
redis==5.0.1
celery==5.3.4
prometheus-client==0.19.0
structlog==23.2.0
